requirements.txt
watchdog
google-generativeai
python-dotenv
pdfminer.six
pikepdf
tqdm
